scripts/hid-research/requirements.txt
# 'hidapi' (not 'hid') bundles hidapi.dll on Windows — same 'import hid' in code
hidapi>=0.14.0
colorama>=0.4.6
